builds/inPyiot/requirements.txt
certifi==2019.3.9
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.8.0
Werkzeug==0.15.2
